requirements.txt
alembic==1.0.5
aniso8601==6.0.0
arrow==0.12.1
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
Flask-UUID==0.2
iso3166==1.0
itsdangerous==1.1.0
Jinja2>=2.10.1
Mako==1.0.8
MarkupSafe==1.1.1
marshmallow==2.16.3
marshmallow-sqlalchemy==0.15.0
more-itertools==6.0.0
phonenumbers==8.10.8
pluggy==0.9.0
psycopg2-binary==2.7.6.1
py==1.8.0
pycountry==18.12.8
pytest==4.2.0
pytest-lazy-fixture==0.5.1
python-dateutil==2.8.0
python-editor==1.0.4
python-slugify==2.0.1
pytz==2018.9
schwifty==2018.9.1
six==1.12.0
SQLAlchemy>=1.3.0
SQLAlchemy-Utils==0.33.8
text-unidecode==1.2
Werkzeug==0.14.1
